Fix bug in practice program. Clean up code. <...>
Fix a bug inside the practice program regarding which lists to
iterate through after calling the ?exit command. No one is going to read
this commit message anyway and I'm tired so fuck explaining what I did.

This commit also cleans up the code a bit and make the output consistent
with regards to newlines etc.